remove old demo, add two simplified examples
