+++ count/empirical_data.cc
@@ -21,6 +21,9 @@
 #include <algorithm>
 #include "count/hll_limits.h"
 
+using libcount::HLL_MIN_PRECISION;
+using libcount::HLL_MAX_PRECISION;
+
 double EMP_alpha(int precision) {
   // TODO(tdial): This could also be a table lookup.
   assert(precision >= HLL_MIN_PRECISION);
